requirements.txt
# MS-REPORT-PY - Microservicio de Reportes
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
openpyxl==3.1.2  # Para exportar Excel

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utils
python-dotenv==1.0.0

